robotframework-rremote/pyproject.toml
[tool.poetry]
name = "robotframework-rremote"
version = "0.1.0"
description = ""
authors = ["Łukasz Sójka <soyacz@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.6"
robotframework = "^3.2.2"
serpent = "^1.30.2"


[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
robotframework-requests = "^0.8.0"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
